workInTemporaryDirectory got an optional parameter 'tempDir'

+++ tests/helpers.py
@@ -45,15 +45,18 @@ if 'SkipTest' not in dir(unittest):
 
 
 @contextmanager
-def workInTemporaryDirectory():
+def workInTemporaryDirectory(tempDir=None):
     """
     Creates a temporary folder to work in. Deletes the folder afterwards.
+
+    :param tempDir: use the given dir as temporary directory. Will not \
+be deleted if given.
     """
-    temporary_folder = tempfile.mkdtemp()
+    temporary_folder = tempDir or tempfile.mkdtemp()
     with cd(temporary_folder):
         yield temporary_folder
 
-    if os.path.exists(temporary_folder):
+    if not tempDir and os.path.exists(temporary_folder):
         shutil.rmtree(temporary_folder)
